@@ -21,8 +21,12 @@
   },
   "scripts": {
     "clean": "rm -rf dist",
-    "start": "npm run dev",
-    "lint": "eslint src",
+    "start": "npm run dev:cloud",
+    "lint": "eslint --ext .js --ext .jsx src",
+    "dev:local": "npm run dev -- --endpoint local",
+    "dev:cloud": "npm run dev -- --endpoint demo",
+    "dev:openfin:local": "node ./node_modules/npm-run-all/bin/npm-run-all --parallel openfin dev:local",
+    "dev:openfin:cloud": "node ./node_modules/npm-run-all/bin/npm-run-all --parallel openfin dev:cloud",
     "dev": "webpack-dev-server --progress --debug --config webpack.config.js --display-error-details",
     "test": "karma start ./karma.conf.js",
     "test:dev": "npm test -- --watch",
@@ -34,34 +38,37 @@
   "dependencies": {
     "animate.css": "3.4.0",
     "autobahn": "0.9.9",
-    "bootstrap-sass": "3.3.6",
     "classnames": "2.2.3",
     "d3": "3.5.12",
     "esp-js": "0.7.1",
-    "esp-js-devtools": "0.0.15",
     "fixed-data-table": "0.6.0",
     "font-awesome": "4.5.0",
     "lodash": "3.10.1",
     "numeral": "1.5.3",
     "nvd3": "1.8.1",
     "react": "0.14.6",
-    "react-dimensions": "1.0.2",
+    "react-dimensions": "1.1.1",
     "react-dom": "0.14.6",
     "react-nvd3": "0.4.1",
-    "react-popout": "0.4.0",
     "react-sparklines": "1.4.2",
     "rx": "4.0.7",
     "ws": "0.8.1"
   },
   "devDependencies": {
+    "esp-js-devtools": "0.0.15",
     "autoprefixer": "6.3.1",
     "autoprefixer-loader": "3.1.0",
-    "babel": "5.8.23",
-    "babel-core": "5.8.33",
+    "babel": "6.5.2",
+    "babel-core": "6.7.6",
     "babel-eslint": "4.1.8",
-    "babel-loader": "5.0.0",
-    "babel-plugin-react-transform": "1.1.1",
-    "babel-runtime": "5.8.20",
+    "babel-loader": "6.2.4",
+    "babel-plugin-react-transform": "2.0.2",
+    "babel-plugin-transform-decorators-legacy": "1.3.4",
+    "babel-plugin-transform-runtime": "6.7.5",
+    "babel-preset-es2015": "6.6.0",
+    "babel-preset-react": "6.5.0",
+    "babel-preset-stage-0": "6.5.0",
+    "babel-runtime": "6.6.1",
     "chalk": "1.1.1",
     "copy-webpack-plugin": "0.3.3",
     "css-loader": "0.23.1",
@@ -73,6 +80,7 @@
     "extract-text-webpack-plugin": "1.0.1",
     "file-loader": "0.8.5",
     "html-webpack-plugin": "1.7.0",
+    "install": "0.6.1",
     "isparta-loader": "2.0.0",
     "jasmine-core": "2.4.1",
     "json-loader": "0.5.4",
@@ -82,16 +90,18 @@
     "karma-sourcemap-loader": "0.3.7",
     "karma-spec-reporter": "0.0.23",
     "karma-webpack": "1.7.0",
+    "minimist": "1.2.0",
     "node-sass": "3.4.2",
-    "openfin-cli": "1.0.1",
+    "npm-run-all": "2.1.0",
+    "openfin-cli": "1.0.2",
     "react-addons-test-utils": "0.14.6",
     "react-transform-catch-errors": "1.0.1",
     "react-transform-hmr": "1.0.1",
     "redbox-react": "1.2.0",
-    "sass-loader": "3.1.2",
+    "sass-loader": "3.2.0",
     "style-loader": "0.13.0",
     "url-loader": "0.5.7",
-    "webpack": "1.12.11",
+    "webpack": "1.13.0",
     "webpack-dev-server": "1.14.1",
     "yargs": "3.31.0"
   }
