@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { Trade, Direction, TradeStatus } from '../model';
+import { Trade, Direction, TradeStatus, TradesUpdate } from '../model';
 import { ReferenceDataService } from '../';
 
 export default class TradeMapper {
@@ -9,13 +9,13 @@ export default class TradeMapper {
     this._referenceDataService = referenceDataService;
   }
 
-  mapFromDtoArray(dtos:Array<Object>) : Array<Trade> {
-    return _.map(
-      dtos,
-      dto => this.mapFromDto(dto));
+  mapFromDto(dto:Object) : Array<Trade> {
+    let trades =  _.map(dto.Trades, trade => this.mapFromTradeDto(trade));
+
+    return new TradesUpdate(dto.IsStateOfTheWorld, dto.IsStale, trades);
   }
 
-  mapFromDto(tradeDto:Object) : Trade {
+  mapFromTradeDto(tradeDto:Object) : Trade {
     let direction = this._mapDirectionFromDto(tradeDto.Direction);
     let status = this._mapTradeStatusFromDto(tradeDto.Status);
     let currencyPair = this._referenceDataService.getCurrencyPair(tradeDto.CurrencyPair);
@@ -29,7 +29,7 @@ export default class TradeMapper {
       tradeDto.SpotRate,
       new Date(tradeDto.TradeDate),
       new Date(tradeDto.ValueDate),
-      status
+      status,
     );
   }
 
