@@ -1,60 +1,164 @@
 import Rx from 'rx';
+import _ from 'lodash';
+import { Trade, TradeNotification } from '../../services/model';
+import { logger } from '../';
+
+const _log:logger.Logger = logger.create('OpenFin');
+
+const REQUEST_LIMIT_CHECK_TOPIC = 'request-limit-check';
 
 export default class OpenFin {
 
-  available:boolean = false;
-  //currentWindow:OpenFinWindow;
-
-  //windows:Array<TearoutWindowInfo> = [];
   tradeClickedSubject:Rx.Subject<string>;
-  //analyticsSubscription:Rx.IDisposable;
-  // pricesSubscription:Rx.IDisposable;
   limitCheckSubscriber:string;
-  requestLimitCheckTopic:string;
-  limitCheckId:number = 1;
+  limitCheckId:number;
 
   constructor() {
     this.tradeClickedSubject = new Rx.Subject();
-    if (typeof fin === 'undefined') return;
-
-    this.available = true;
     this.limitCheckId = 1;
-    this.requestLimitCheckTopic = 'request-limit-check';
-    console.log('Application is running in OpenFin container');
-    // this.setToolbarAsDraggable();
+    this.limitCheckSubscriber = null;
+    if (this.isRunningInOpenFin) {
+      this._initializeLimitChecker();
+    }
+  }
+
+  get isRunningInOpenFin() {
+    return typeof fin !== 'undefined';
+  }
+
+  close(window = this._currentWindow){
+    window.close(true, () => _log.info('Window closed with success.'), err => _log.error('Failed to close window.', err));
+  }
+
+  minimize(window = this._currentWindow){
+    window.minimize(() => _log.info('Window minimized with success.'), err => _log.error('Failed to minimize window.', err));
+  }
+
+  maximize(window = this._currentWindow){
+    window.getState(state => {
+        switch (state){
+          case 'maximized':
+          case 'restored':
+          case 'minimized':
+            window.restore(() => window.bringToFront());
+            break;
+          default:
+            window.maximize(() => _log.info('Window maximized with success.'), err => _log.error('Failed to maximize window.', err));
+        }
+      });
   }
 
   checkLimit(executablePrice, notional:number, tradedCurrencyPair:string):Rx.Observable<boolean> {
+    let _this = this;
     return Rx.Observable.create(observer => {
         let disposables = new Rx.CompositeDisposable();
-        if (!this.available || this.limitCheckSubscriber == null) {
-          console.log('client side limit check not up, will delegate to to server');
+        if (_this.limitCheckSubscriber === null) {
+          _log.debug('client side limit check not up, will delegate to to server');
           observer.onNext(true);
           observer.onCompleted();
         } else {
-          console.log('checking if limit is ok with ' + this.limitCheckSubscriber);
-          var topic = 'limit-check-response' + (this.limitCheckId++);
-          var limitCheckResponse:(msg:any) => void = (msg) => {
-            console.log(this.limitCheckSubscriber + ' limit check response was ' + msg);
+          _log.debug(`checking if limit is ok with ${_this.limitCheckSubscriber}`);
+          const topic = `limit-check-response (${_this.limitCheckId++})`;
+          let limitCheckResponse:(msg:any) => void = (msg) => {
+            _log.debug(`${_this.limitCheckSubscriber} limit check response was ${msg}`);
             observer.onNext(msg.result);
             observer.onCompleted();
           };
 
-          fin.desktop.InterApplicationBus.subscribe(this.limitCheckSubscriber, topic, limitCheckResponse);
+          fin.desktop.InterApplicationBus.subscribe(_this.limitCheckSubscriber, topic, limitCheckResponse);
 
-          fin.desktop.InterApplicationBus.send(this.limitCheckSubscriber, this.requestLimitCheckTopic, {
-            id: this.limitCheckId,
+          fin.desktop.InterApplicationBus.send(_this.limitCheckSubscriber, REQUEST_LIMIT_CHECK_TOPIC, {
+            id: _this.limitCheckId,
             responseTopic: topic,
             tradedCurrencyPair: tradedCurrencyPair,
             notional: notional,
-            rate: executablePrice.rate
+            rate: executablePrice
           });
 
           disposables.add(Rx.Disposable.create(() => {
-            fin.desktop.InterApplicationBus.unsubscribe(this.limitCheckSubscriber, topic, limitCheckResponse);
+            fin.desktop.InterApplicationBus.unsubscribe(_this.limitCheckSubscriber, topic, limitCheckResponse);
           }));
         }
         return disposables;
       });
+  }
+
+  get _currentWindow() {
+    return fin.desktop.Window.getCurrent();
+  }
+
+  displayCurrencyChart(symbol){
+    let chartIqAppId = 'ChartIQ';
+    fin.desktop.System.getAllApplications((apps) => {
+      let chartIqApp = _.find(apps, ((app) => {
+        return app.isRunning && app.uuid === chartIqAppId;
+      }));
+      if(chartIqApp) {
+        this._refreshCurrencyChart(symbol);
+      } else {
+        this._launchCurrencyChart(symbol);
+      }
+    });
+  }
+
+  /**
+   * Initialize limit checker
+   * @private
+   */
+  _initializeLimitChecker() {
+    fin.desktop.main(() => {
+      fin.desktop.InterApplicationBus.addSubscribeListener(function(uuid, topic) {
+        if (topic === REQUEST_LIMIT_CHECK_TOPIC) {
+          _log.info(`${uuid} has subscribed as a limit checker`);
+          // There will only be one. If there are more, last subscriber will be used
+          this.limitCheckSubscriber = uuid;
+        }
+      }.bind(this));
+      fin.desktop.InterApplicationBus.addUnsubscribeListener(function(uuid, topic) {
+        if (topic === REQUEST_LIMIT_CHECK_TOPIC) {
+          _log.info(`${uuid} has unsubscribed as a limit checker`);
+          this.limitCheckSubscriber = null;
+        }
+      }.bind(this));
+    });
+  }
+
+  _refreshCurrencyChart(symbol){
+    let interval = 5;
+    let chartIqAppId = 'ChartIQ';
+    fin.desktop.InterApplicationBus.publish('chartiq:main:change_symbol', { symbol: symbol, interval: interval });
+  }
+
+  _launchCurrencyChart(symbol){
+    let interval = 5;
+    let chartIqAppId = 'ChartIQ';
+    let url = `http://openfin.chartiq.com/0.5/chartiq-shim.html?symbol=${symbol}&period=${interval}`;
+    let name = `chartiq_${(new Date()).getTime()}`;
+    const applicationIcon = 'http://openfin.chartiq.com/0.5/img/openfin-logo.png';
+    let app = new fin.desktop.Application({
+      uuid: chartIqAppId,
+      url: url,
+      name: name,
+      applicationIcon: applicationIcon,
+      mainWindowOptions:{
+        autoShow: false
+      }
+    }, function(){
+      app.run();
+    });
+  }
+
+  openTradeNotification(trade:Trade): void {
+    if (!this.isRunningInOpenFin) return;
+
+    let tradeNotification = new TradeNotification(trade);
+
+    let notification = new fin.desktop.Notification({
+      url: '/notification.html',
+      message: tradeNotification,
+      onMessage: () => {
+        this.maximise();
+      }
+    });
   }
 }
