@@ -9,11 +9,23 @@ export default class AutobahnConnectionProxy {
   connection:autobahn.Connection;
 
   constructor(url:string, realm:string) {
+    const useSecure = location.protocol === 'https:';
+    const securePort = 8000;
+    const defaultPort = 8080;
     this.connection = new autobahn.Connection({
-      url: url,
       realm: realm,
       use_es6_promises: true,
-      max_retries: -1 // unlimited retries
+      max_retries: -1, // unlimited retries,
+      transports: [
+        {
+          type: 'websocket',
+          url: useSecure ? `wss://${url}:${securePort}/ws` : `ws://${url}:${defaultPort}/ws`
+        },
+        {
+          type: 'longpoll',
+          url: useSecure ? `https://${url}:${securePort}/lp` : `http://${url}:${defaultPort}/lp`
+        }
+      ]
     });
   }
 
