@@ -1,2 +1,2 @@
 export { default as Modal } from './modal/modal.jsx';
-export { default as PageContainer } from './pageContainer/pageContainer.jsx';
+export { default as Chrome } from './chrome/views/chrome.jsx';
