@@ -6,7 +6,7 @@ var _log:logger.Logger = logger.create('ModelBase');
 export default class ModelBase extends DisposableBase {
   _modelId:string;
   router:Router;
-
+  
   constructor(modelId, router) {
     super();
     Guard.isString(modelId, 'modelId required and must be a string');
@@ -28,7 +28,7 @@ export default class ModelBase extends DisposableBase {
   /**
    * Runs the given action on the dispatch loop for this model, ensures that any model observer will be notified of the change
    * @param action
-     */
+   */
   ensureOnDispatchLoop(action:() => void) {
     // TODO update when https://github.com/esp/esp-js/issues/86 is implemented
     this.router.runAction(this.modelId, ()=>{
