@@ -1,10 +1,7 @@
-import _ from 'lodash';
 import React from 'react';
 import { ViewBase } from '../../../common';
-import { PageContainer } from '../../../common/components';
 import { RegionModel, RegionModelRegistration } from '../../model';
 import { createViewForModel } from '../../';
-import { router } from '../../../../system';
 
 export default class SingleItemRegionView extends ViewBase {
   constructor() {
@@ -21,6 +18,6 @@ export default class SingleItemRegionView extends ViewBase {
     }
     let modelRegistration:RegionModelRegistration = model.modelRegistrations[0];
     let view = createViewForModel(modelRegistration.model, modelRegistration.displayContext);
-    return ( <div>{view}</div>)
+    return view;
   }
 }
