@@ -3,9 +3,13 @@ import './workspaceItemContainer.scss';
 
 export default class WorkspaceItemContainer extends React.Component {
 
+  static propTypes = {
+    children: React.PropTypes.element.isRequired
+  };
+
   render(){
     return (
-      <div>
+      <div className='workspace-region__item'>
         {this.props.children}
       </div>);
   }
