@@ -4,6 +4,8 @@ import { SpotTileModel } from './model';
 import { PricingService, ExecutionService } from '../../services';
 import { CurrencyPair } from '../../services/model';
 import { RegionManager, RegionNames } from '../regions';
+import { SchedulerService, } from '../../system';
+import { OpenFin } from '../../system/openFin';
 
 let modelIdKey = 1;
 
@@ -17,17 +19,23 @@ export default class SpotTileFactory {
   _pricingService:PricingService;
   _executionService:ExecutionService;
   _regionManager:RegionManager;
+  _schedulerService:SchedulerService;
+  _openFin:OpenFin;
 
   constructor(
     router:Router,
     pricingService:PricingService,
     executionService:ExecutionService,
-    regionManager:RegionManager
+    regionManager:RegionManager,
+    schedulerService: SchedulerService,
+    openFin: OpenFin
   ) {
     this._router = router;
     this._pricingService = pricingService;
-    this._executionService =executionService;
+    this._executionService = executionService;
     this._regionManager = regionManager;
+    this._schedulerService = schedulerService;
+    this._openFin = openFin;
   }
 
   createTileModel(currencyPair:CurrencyPair) {
@@ -37,7 +45,9 @@ export default class SpotTileFactory {
       this._router,
       this._pricingService,
       this._executionService,
-      this._regionManager
+      this._regionManager,
+      this._schedulerService,
+      this._openFin
     );
     spotTileModel.observeEvents();
     return spotTileModel;
