@@ -1,30 +1,47 @@
 import React from 'react';
 import { TradeExecutionNotification } from '../model';
+import { TradeStatus } from '../../../services/model';
+import classnames from 'classnames';
+import './tradeNotification.scss';
 
 export default class TradeNotification extends React.Component {
 
   static propTypes = {
+    className: React.PropTypes.string,
     tradeExecutionNotification: React.PropTypes.instanceOf(TradeExecutionNotification).isRequired,
     onDismissedClicked: React.PropTypes.func
   };
 
   render(){
-    let { tradeExecutionNotification, onDismissedClicked } = this.props;
+    let { tradeExecutionNotification, onDismissedClicked, className } = this.props;
 
     if(tradeExecutionNotification.hasError) {
-      return (<div className='blocked summary-state animated flipInX'>
-        <span className='key'>Error:</span> {tradeExecutionNotification.error}. The execution status is unknown. Please contact your sales rep.
-        <a href='#' className='pull-right dismiss-message' onClick={onDismissedClicked}>Done</a>
+      let classNames = classnames(
+        'trade-notification',
+        'trade-notification--error',
+        className
+      );
+      return (<div className={classNames}>
+        {tradeExecutionNotification.error}. The execution status is unknown. Please contact your sales rep.
+        <a href='#' className='trade-notification__button--dismiss' onClick={onDismissedClicked}>Done</a>
       </div>);
     } else {
+      let classNames = classnames(
+        'trade-notification',
+        className,
+        {
+          'trade-notification--rejected': tradeExecutionNotification.status === TradeStatus.Rejected
+        }
+      );
       return (
-        <div className={tradeExecutionNotification.status + ' summary-state animated flipInX'}>
-          <span className='key'>{tradeExecutionNotification.direction}</span> {tradeExecutionNotification.dealtCurrency} {tradeExecutionNotification.notional}<br/>
-          <span className='key'>vs</span> {tradeExecutionNotification.termsCurrency}
-          <span className='key'>at </span> {tradeExecutionNotification.spotRate}<br/>
-          <span className='key'>{tradeExecutionNotification.formattedValueDate}</span><br/>
-          <span className='key'>Trade ID</span> {tradeExecutionNotification.tradeId}
-          <a href='#' className='pull-right dismiss-message' onClick={onDismissedClicked}>{tradeExecutionNotification.status}</a>
+        <div className={classNames}>
+          <span className='trade-notification__trade-status'>{tradeExecutionNotification.status.name}</span>
+          <span className='trade-notification__label'>{tradeExecutionNotification.direction} </span><span className='trade-notification__value'>{tradeExecutionNotification.dealtCurrency} {tradeExecutionNotification.notional}</span><br/>
+          <span className='trade-notification__label'> vs </span><span className='trade-notification__value'>{tradeExecutionNotification.termsCurrency}</span>
+          <span className='trade-notification__label'> at </span><span className='trade-notification__value'>{tradeExecutionNotification.spotRate}</span><br/>
+          <span className='trade-notification__value'>{tradeExecutionNotification.formattedValueDate}</span><br/>
+          <span className='trade-notification__label'>Trade ID </span><span className='trade-notification__value'>{tradeExecutionNotification.tradeId}</span>
+          <a href='#' className='trade-notification__button--dismiss' onClick={onDismissedClicked}>Done</a>
         </div>
       );
     }
