@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Reactive.Concurrency;
 using System.Text;
 using System.Threading.Tasks;
 using Adaptive.ReactiveTrader.Messaging.Abstraction;
@@ -15,6 +16,7 @@ namespace Adaptive.ReactiveTrader.Messaging.WAMP
     {
         protected static readonly ILog Log = LogManager.GetLogger<RpcOperation>();
 
+        private readonly IScheduler _scheduler = TaskPoolScheduler.Default;
         private readonly Func<IRequestContext, IMessage, Task<TResponse>> _serviceMethod;
 
         public RpcResponseOperation(string name, Func<IRequestContext, IMessage, Task<TResponse>> serviceMethod)
@@ -34,7 +36,7 @@ namespace Adaptive.ReactiveTrader.Messaging.WAMP
             caller.Error(WampObjectFormatter.Value,
                          dummyDetails,
                          "wamp.error.runtime_error",
-                         new object[] {"Expected parameters"});
+                         new object[] { "Expected parameters" });
         }
 
         public void Invoke<TMessage>(IWampRawRpcOperationRouterCallback caller,
@@ -54,7 +56,7 @@ namespace Adaptive.ReactiveTrader.Messaging.WAMP
             InnerInvoke(_serviceMethod, caller, formatter, arguments);
         }
 
-        private static void InnerInvoke<T>(Func<IRequestContext, IMessage, Task<TResponse>> serviceMethod,
+        private void InnerInvoke<T>(Func<IRequestContext, IMessage, Task<TResponse>> serviceMethod,
                                            IWampRawRpcOperationRouterCallback caller,
                                            IWampFormatter<T> formatter,
                                            T[] arguments)
@@ -63,31 +65,45 @@ namespace Adaptive.ReactiveTrader.Messaging.WAMP
 
             try
             {
-                var x = formatter.Deserialize<MessageDto>(arguments[0]);
-
-                var message = new Message
+                _scheduler.Schedule(async () =>
                 {
-                    ReplyTo = new WampTransientDestination(x.ReplyTo),
-                    Payload = Encoding.UTF8.GetBytes(x.Payload.ToString()) // TODO need to stop this from deserializing
-                };
+                    try
+                    {
+                        var x = formatter.Deserialize<MessageDto>(arguments[0]);
 
-                var userSession = new UserSession
-                {
-                    Username = x.Username
-                };
+                        var message = new Message
+                        {
+                            ReplyTo = new WampTransientDestination(x.ReplyTo),
+                            Payload = Encoding.UTF8.GetBytes(x.Payload.ToString()) // TODO need to stop this from deserializing
+                        };
 
-                var userContext = new RequestContext(message, userSession);
+                        var userSession = new UserSession
+                        {
+                            Username = x.Username
+                        };
 
-                var response = serviceMethod(userContext, message).Result;
+                        var userContext = new RequestContext(message, userSession);
 
+                        var response = await serviceMethod(userContext, message);
 
-                caller.Result(WampObjectFormatter.Value, dummyDetails, new object[] {response});
+                        caller.Result(WampObjectFormatter.Value, dummyDetails, new object[] { response });
+                    }
+                    catch (Exception e1)
+                    {
+                        OnError(caller, e1, dummyDetails);
+                    }
+                });
             }
-            catch (Exception e)
+            catch (Exception e2)
             {
-                Log.Error(e);
-                caller.Error(WampObjectFormatter.Value, dummyDetails, e.Message);
+                OnError(caller, e2, dummyDetails);
             }
+        }
+
+        private static void OnError(IWampRawRpcOperationRouterCallback caller, Exception exception, YieldOptions dummyDetails)
+        {
+            Log.Error(exception);
+            caller.Error(WampObjectFormatter.Value, dummyDetails, exception.Message);
         }
     }
 }
